chore: move to pbr material
