+++ packages/core/src/material/PBRBaseMaterial.ts
@@ -1,4 +1,4 @@
-import { Color, Vector2, Vector3, Vector4 } from "@galacean/engine-math";
+import { Color, Vector4 } from "@galacean/engine-math";
 import { Logger, ShaderProperty } from "..";
 import { Engine } from "../Engine";
 import { Shader } from "../shader/Shader";
@@ -19,10 +19,6 @@ export abstract class PBRBaseMaterial extends BaseMaterial {
   private static _clearCoatRoughnessProp = ShaderProperty.getByName("material_ClearCoatRoughness");
   private static _clearCoatRoughnessTextureProp = ShaderProperty.getByName("material_ClearCoatRoughnessTexture");
   private static _clearCoatNormalTextureProp = ShaderProperty.getByName("material_ClearCoatNormalTexture");
-
-  private static _anisotropyInfoProp = ShaderProperty.getByName("material_AnisotropyInfo");
-  private static _anisotropyTextureProp = ShaderProperty.getByName("material_AnisotropyTexture");
-  private _anisotropyDirection = new Vector2();
 
   /**
    * Base color.
@@ -248,57 +244,6 @@ export abstract class PBRBaseMaterial extends BaseMaterial {
   }
 
   /**
-   * The Amount of anisotropy. Scalar between −1 and 1.
-   */
-  get anisotropy(): number {
-    return this.shaderData.getVector3(PBRBaseMaterial._anisotropyInfoProp).z;
-  }
-
-  set anisotropy(value: number) {
-    const anisotropyInfo = this.shaderData.getVector3(PBRBaseMaterial._anisotropyInfoProp);
-    if (!!anisotropyInfo.z !== !!value) {
-      if (value === 0) {
-        this.shaderData.disableMacro("MATERIAL_ENABLE_ANISOTROPY");
-      } else {
-        this.shaderData.enableMacro("MATERIAL_ENABLE_ANISOTROPY");
-      }
-    }
-    anisotropyInfo.z = value;
-  }
-
-  /**
-   * The direction of the surface to control the shape of the specular highlights
-   */
-  get anisotropyDirection(): Vector2 {
-    this._anisotropyDirection.copyFrom(this.shaderData.getVector3(PBRBaseMaterial._anisotropyInfoProp));
-
-    return this._anisotropyDirection;
-  }
-
-  set anisotropyDirection(value: Vector2) {
-    if (value !== this._anisotropyDirection) {
-      this._anisotropyDirection.copyFrom(value);
-    }
-  }
-
-  /**
-   * The anisotropy texture.
-   */
-  get anisotropyTexture(): Texture2D {
-    return <Texture2D>this.shaderData.getTexture(PBRBaseMaterial._anisotropyTextureProp);
-  }
-
-  set anisotropyTexture(value: Texture2D) {
-    this.shaderData.setTexture(PBRBaseMaterial._anisotropyTextureProp, value);
-
-    if (value) {
-      this.shaderData.enableMacro("MATERIAL_HAS_ANISOTROPY_TEXTURE");
-    } else {
-      this.shaderData.disableMacro("MATERIAL_HAS_ANISOTROPY_TEXTURE");
-    }
-  }
-
-  /**
    * Create a pbr base material instance.
    * @param engine - Engine to which the material belongs
    * @param shader - Shader used by the material
@@ -321,15 +266,5 @@ export abstract class PBRBaseMaterial extends BaseMaterial {
 
     shaderData.setFloat(PBRBaseMaterial._clearCoatProp, 0);
     shaderData.setFloat(PBRBaseMaterial._clearCoatRoughnessProp, 0);
-
-    shaderData.setVector3(PBRBaseMaterial._anisotropyInfoProp, new Vector3(1, 0, 0));
-    // @ts-ignore
-    this._anisotropyDirection._onValueChanged = this._onAnisotropyDirectionChanged.bind(this);
-  }
-
-  private _onAnisotropyDirectionChanged(): void {
-    const anisotropyInfo = this.shaderData.getVector3(PBRBaseMaterial._anisotropyInfoProp);
-    anisotropyInfo.x = this._anisotropyDirection.x;
-    anisotropyInfo.y = this._anisotropyDirection.y;
   }
 }
